requirements.txt
# VPN Manager Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Telegram Bot
python-telegram-bot>=20.7

# Security
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0

# SSH Remote Management
paramiko>=3.0.0

# Utilities
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
python-multipart>=0.0.6
aiofiles>=23.2.0
httpx>=0.25.0
requests>=2.31.0

# Monitoring & Logging
loguru>=0.7.2
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Client Portal & Crypto Payments
aiocryptopay>=0.4.8
bcrypt>=4.0.0
PyJWT>=2.8.0
email-validator>=2.0.0

# Payment Plugins (pre-installed, activated via .env keys)
stripe>=9.0.0
mollie-api-python>=3.0.0
razorpay>=1.4.0
payme-pkg>=3.0.0
